+++ src/renderer/renderScene.ts
@@ -420,7 +420,9 @@ export const renderScene = (
   // Paint remote pointers
   for (const clientId in sceneState.remotePointerViewportCoords) {
     let { x, y } = sceneState.remotePointerViewportCoords[clientId];
-    const username = sceneState.remotePointerUsernames[clientId];
+
+    x -= appState.offsetLeft;
+    y -= appState.offsetTop;
 
     const width = 9;
     const height = 14;
@@ -474,6 +476,8 @@ export const renderScene = (
     context.lineTo(x, y);
     context.fill();
     context.stroke();
+
+    const username = sceneState.remotePointerUsernames[clientId];
 
     if (!isOutOfBounds && username) {
       const offsetX = x + width;